requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
streamlit==1.34.0
chromadb==0.4.24
langchain==0.1.17
langchain-community==0.0.36
pypdf==4.2.0
python-dotenv==1.0.1
requests==2.32.2
uuid==1.30
shutilwhich==1.1.0
google-generativeai==0.5.4
ollama
watchdog==4.0.0
python-multipart==0.0.9